+++ warlock_manager/libs/get_wan_ip.py
@@ -1,18 +1,7 @@
-from urllib import request
-from urllib import error as urllib_error
+from typing_extensions import deprecated
+from warlock_manager.libs.ip import get_wan_ip as new_get_wan_ip
 
 
+@deprecated('This functionality has moved to warlock_manager.libs.ip.get_wan_ip')
 def get_wan_ip() -> str | None:
-	"""
-	Get the external IP address of this server
-	:return: str: The external IP address as a string, or None if it cannot be determined
-	"""
-	try:
-		with request.urlopen('http://wan.eval.bz', timeout=2) as resp:
-			return resp.read().decode('utf-8')
-	except (urllib_error.HTTPError, urllib_error.URLError, TimeoutError):
-		try:
-			with request.urlopen('https://api.ipify.org', timeout=2) as resp:
-				return resp.read().decode('utf-8')
-		except (urllib_error.HTTPError, urllib_error.URLError, TimeoutError):
-			return None
+	return new_get_wan_ip()